Updating Fitbit Food Get Uri

+++ src/Biotrackr.Food.Svc/Biotrackr.Food.Svc/Models/FitbitEntities/NutritionalValues.cs
@@ -8,11 +8,11 @@ namespace Biotrackr.Food.Svc.Models.FitbitEntities
 {
     public class NutritionalValues
     {
-        public int calories { get; set; }
-        public int carbs { get; set; }
+        public double calories { get; set; }
+        public double carbs { get; set; }
         public double fat { get; set; }
-        public int fiber { get; set; }
-        public int protein { get; set; }
-        public int sodium { get; set; }
+        public double fiber { get; set; }
+        public double protein { get; set; }
+        public double sodium { get; set; }
     }
 }
